+++ Changelog.md
@@ -11,6 +11,7 @@ Breaking changes:
 
 Language Features:
  * Allow global enums and structs.
+ * Allow explicit conversions from ``address`` to ``address payable`` via ``payable(...)``.
 
 
 Compiler Features: